Put pointless prefix items back in pointless menu
The information architecture and menu flow of this breaks otherwise.
I woud kind of like to rearrange this more but the menu highlighting
logic here depends on the prefix, so things like Climbing per Mile
really ought to be in Data Exploration but the routes for them would
also have to change.

I'll create an issue to cover this in more detail.